ikiAI_Agents/agents/survey_analysis_agent.py
from backend.config import config
from backend.database import ChromaDB

class SurveyAnalysisAgent:
    def __init__(self):
        self.db = ChromaDB(config)

    def analyze_and_store_goals(self, survey_data):
        # Analyze survey data and extract goals
        goals = self.analyze_survey_data(survey_data)
        
        # Store goals in database
        self.db.store_goals(goals)

    def analyze_survey_data(self, survey_data):
        # Placeholder for survey data analysis logic
        goals = []
        return goals

    def generate_daily_tasks(self, goals):
        # Placeholder for daily task generation logic
        tasks = []
        return tasks

    def calculate_task_completion_percentage(self, user_id):
        # Retrieve user's tasks from database
        tasks = self.db.retrieve_tasks(user_id)

        # Calculate task completion percentage
        completed_tasks = [task for task in tasks if task['status'] == 'completed']
        completion_percentage = len(completed_tasks) / len(tasks) * 100

        return completion_percentage
